scripts/requirements.txt
gradio>=4.0.0
openai
langchain-openai
langgraph
pydantic
typing-extensions
python-dotenv
requests
replicate
